[#93] Moved `stalker.db.session.DBSession` test to its own test file.
